fix bugs

+++ app/src/main/java/com/example/group_11_project_app_seg2105/tutor/AvailabilityAdapter.java
@@ -25,7 +25,7 @@ public class AvailabilityAdapter extends RecyclerView.Adapter<AvailabilityAdapte
     private final DeleteCallback deleteCallback;  // NEW CALLBACK for safe delete
 
     public interface DeleteCallback {
-        void onRequestDelete(int slotId);
+        void onRequestDelete(long slotId);
     }
 
     public AvailabilityAdapter(
